rename metrics repo module
