+++ test/integration/pipelines/tuning/test_tuner_builder.py
@@ -16,7 +16,7 @@ from fedot.core.optimisers.objective.data_source_splitter import DataSourceSplit
 from fedot.core.optimisers.objective.metrics_objective import MetricsObjective
 from fedot.core.pipelines.tuning.search_space import PipelineSearchSpace
 from fedot.core.pipelines.tuning.tuner_builder import TunerBuilder
-from fedot.core.repository.quality_metrics_repository import ClassificationMetricsEnum, MetricIDType
+from fedot.core.repository.metrics_repository import ClassificationMetricsEnum, MetricIDType
 from test.integration.pipelines.tuning.test_pipeline_tuning import get_not_default_search_space
 from test.unit.optimizer.test_pipeline_objective_eval import pipeline_first_test
 from test.unit.validation.test_table_cv import get_classification_data
